+++ base/documentation.py
@@ -43,6 +43,7 @@ base.VERSION.changed("1.12.1", "Extended CONTRIBUTING.md in `base.documentation`
 base.VERSION.fixed("1.12.2", "Typos in `base.documentation` ")
 base.VERSION.added("1.12.2", "Step to merge request instructions in `base.documentation` ")
 base.VERSION.changed("1.12.5", "Enhanced contribution documentation in `base.documentation` ")
+base.VERSION.changed("1.12.6", "Refactored documentation of class members into own function in `base.documentation` ")
 
 
 def write_changelog(name: str, version_history: base.VersionCollection, file_path: str) -> None:
@@ -90,11 +91,15 @@ def document_components(components_module: types.ModuleType, file_path: str) -> 
     with open(file_path, "w", encoding="utf-8") as f:
         f.write("# Components\n")
         f.write("This file lists all components that are currently included in the Landscape Model core.\n")
-        f.write(f"It was automatically created on {datetime.date.today()}.\n")
-        for name, member in components_module.__dict__.items():
-            if inspect.isclass(member) and issubclass(member, base.Component):
-                f.write(f"\n\n## {name}")
-                f.write(member.__doc__)
+        _document_member(components_module, f)
+
+
+def _document_member(components_module: types.ModuleType, f: typing.TextIO) -> None:
+    f.write(f"It was automatically created on {datetime.date.today()}.\n")
+    for name, member in components_module.__dict__.items():
+        if inspect.isclass(member) and issubclass(member, base.Component):
+            f.write(f"\n\n## {name}")
+            f.write(member.__doc__)
 
 
 def document_observers(observers_module: types.ModuleType, file_path: str) -> None:
@@ -111,11 +116,7 @@ def document_observers(observers_module: types.ModuleType, file_path: str) -> No
     with open(file_path, "w", encoding="utf-8") as f:
         f.write("# Observers\n")
         f.write("This file lists all observers that are currently included in the Landscape Model core.\n")
-        f.write(f"It was automatically created on {datetime.date.today()}.\n")
-        for name, member in observers_module.__dict__.items():
-            if inspect.isclass(member) and issubclass(member, base.Observer):
-                f.write(f"\n\n## {name}")
-                f.write(member.__doc__)
+        _document_member(observers_module, f)
 
 
 def document_stores(stores_module: types.ModuleType, file_path: str) -> None:
@@ -132,11 +133,7 @@ def document_stores(stores_module: types.ModuleType, file_path: str) -> None:
     with open(file_path, "w", encoding="utf-8") as f:
         f.write("# Stores\n")
         f.write("This file lists all stores that are currently included in the Landscape Model core.\n")
-        f.write(f"It was automatically created on {datetime.date.today()}.\n")
-        for name, member in stores_module.__dict__.items():
-            if inspect.isclass(member) and issubclass(member, base.Store):
-                f.write(f"\n\n## {name}")
-                f.write(member.__doc__)
+        _document_member(stores_module, f)
 
 
 def document_component(